Rename config file to mqtt_agent_demo_config.h
