+++ demos/coreMQTT_Agent/simple_sub_pub_demo.c
@@ -52,7 +52,7 @@
 #include "task.h"
 
 /* Demo Specific configs. */
-#include "mqtt_demo_connection_sharing_config.h"
+#include "mqtt_agent_demo_config.h"
 
 /* MQTT library includes. */
 #include "core_mqtt.h"
