+++ projects/infineon/xmc4800_plus_optiga_trust_x/dave4/aws_demos/.project
@@ -2064,9 +2064,9 @@
 			<locationURI>AFR_HOME/vendors/infineon/boards/xmc4800_plus_optiga_trust_x/aws_demos/config_files/jobs_demo_config.h</locationURI>
 		</link>
 		<link>
-			<name>config_files/mqtt_demo_connection_sharing_config.h</name>
+			<name>config_files/mqtt_agent_demo_config.h</name>
 			<type>1</type>
-			<locationURI>AFR_HOME/vendors/infineon/boards/xmc4800_plus_optiga_trust_x/aws_demos/config_files/mqtt_demo_connection_sharing_config.h</locationURI>
+			<locationURI>AFR_HOME/vendors/infineon/boards/xmc4800_plus_optiga_trust_x/aws_demos/config_files/mqtt_agent_demo_config.h</locationURI>
 		</link>
 		<link>
 			<name>config_files/mqtt_demo_mutual_auth_config.h</name>
