+++ projects/mediatek/mt7697hx-dev-kit/uvision/aws_demos/aws_demos.uvprojx
@@ -2380,9 +2380,9 @@
 							<FilePath>../../../../../vendors/mediatek/boards/mt7697hx-dev-kit/aws_demos/config_files/lwipopts.h</FilePath>
 						</File>
 						<File>
-							<FileName>mqtt_demo_connection_sharing_config.h</FileName>
+							<FileName>mqtt_agent_demo_config.h</FileName>
 							<FileType>5</FileType>
-							<FilePath>../../../../../vendors/mediatek/boards/mt7697hx-dev-kit/aws_demos/config_files/mqtt_demo_connection_sharing_config.h</FilePath>
+							<FilePath>../../../../../vendors/mediatek/boards/mt7697hx-dev-kit/aws_demos/config_files/mqtt_agent_demo_config.h</FilePath>
 						</File>
 						<File>
 							<FileName>mqtt_demo_mutual_auth_config.h</FileName>
