+++ projects/nuvoton/numaker_iot_m487_wifi/uvision/aws_demos/aws_demos.uvproj
@@ -2504,9 +2504,9 @@
 							<FilePath>../../../../../vendors/nuvoton/boards/numaker_iot_m487_wifi/aws_demos/config_files/jobs_demo_config.h</FilePath>
 						</File>
 						<File>
-							<FileName>mqtt_demo_connection_sharing_config.h</FileName>
+							<FileName>mqtt_agent_demo_config.h</FileName>
 							<FileType>5</FileType>
-							<FilePath>../../../../../vendors/nuvoton/boards/numaker_iot_m487_wifi/aws_demos/config_files/mqtt_demo_connection_sharing_config.h</FilePath>
+							<FilePath>../../../../../vendors/nuvoton/boards/numaker_iot_m487_wifi/aws_demos/config_files/mqtt_agent_demo_config.h</FilePath>
 						</File>
 						<File>
 							<FileName>mqtt_demo_mutual_auth_config.h</FileName>
