+++ vendors/cypress/boards/CY8CKIT_064S0S2_4343W/aws_demos/config_files/mqtt_agent_demo_config.h
