+++ vendors/infineon/boards/xmc4800_iotkit/aws_demos/config_files/mqtt_agent_demo_config.h
