+++ vendors/infineon/boards/xmc4800_plus_optiga_trust_x/aws_demos/config_files/mqtt_agent_demo_config.h
