+++ vendors/mediatek/boards/mt7697hx-dev-kit/aws_demos/config_files/mqtt_agent_demo_config.h
