+++ vendors/ti/boards/cc3220_launchpad/aws_demos/config_files/mqtt_agent_demo_config.h
